Fix macOS build around hunsepll dictionaries. Correct spelling in readme.txt

+++ Builder/import-hunspell.sh
@@ -6,6 +6,13 @@
 # dictionaries for hunspell:
 #   https://sourceforge.net/projects/wordlist/files/speller/
 
+if [ "${#}" -lt 2 ]
+then
+    colour-print "<>error Error:<> too few args"
+    echo "Example: ./import-hunspell.sh /Volumes/Downloads/BEmacs/dependencies/hunspell-1.7.0.tar.gz /Volumes/Downloads/BEmacs/dependencies/*.zip"
+    exit 1
+fi
+
 colour-print "<>info Info:<> install hunspell source"
 rm -rf ../Imports/hunspell-*
 mkdir -p ../Imports
@@ -16,20 +23,21 @@ shift
 
 ln -sf hunspell-* hunspell
 cd hunspell
+rm -rf tmp.dict
+mkdir tmp.dict
 
 colour-print "<>info Info:<> installing dictionary kits"
 for DICT in "$@"
 do
-    rm -rf tmp.dict
-    mkdir tmp.dict
     unzip -q ${DICT} -d tmp.dict
     for FILE in tmp.dict/*.dic tmp.dict/*.aff
     do
         if [ -e "$FILE" ]
         then
-            colour-print "<>info Info:<> Dictionary $FILE"
-            NEWFILE=$(basename $FILE)
-            NEWFILE=${NEWFILE/-large/}
+            NEWFILE=${FILE/-large/}
+            mv "$FILE" "$NEWFILE"
+            colour-print "<>info Info:<> Dictionary $PWD/$NEWFILE"
         fi
     done
 done
+
